For #12069 - Break the spocs functionality from PocketStoriesUseCases to SpocsUseCases
